+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/executor/handler/error/JobExceptionHandler.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.executor.handler;
+package org.apache.shardingsphere.elasticjob.lite.executor.handler.error;
 
 /**
  * Job exception handler.
