+++ elastic-job-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/executor/handler/threadpool/JobExecutorServiceHandler.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,25 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.executor.handler.impl;
-
-import org.apache.shardingsphere.elasticjob.lite.executor.handler.JobExecutorServiceHandler;
-import org.apache.shardingsphere.elasticjob.lite.util.concurrent.ElasticJobExecutorService;
+package org.apache.shardingsphere.elasticjob.lite.executor.handler.threadpool;
 
 import java.util.concurrent.ExecutorService;
 
 /**
- * Default job executor service handler.
+ * Job executor service handler.
  */
-public final class DefaultJobExecutorServiceHandler implements JobExecutorServiceHandler {
+public interface JobExecutorServiceHandler {
     
-    @Override
-    public ExecutorService createExecutorService(final String jobName) {
-        return new ElasticJobExecutorService("inner-job-" + jobName, Runtime.getRuntime().availableProcessors() * 2).createExecutorService();
-    }
+    /**
+     * Create executor service.
+     * 
+     * @param jobName job name
+     * 
+     * @return executor service
+     */
+    ExecutorService createExecutorService(String jobName);
     
-    @Override
-    public String getType() {
-        return "DEFAULT";
-    }
+    /**
+     * Get job executor service handler type.
+     * 
+     * @return job executor service handler type
+     */
+    String getType();
 }
