+++ elastic-job-lite-core/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.executor.handler.error.JobExceptionHandler
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.executor.handler.impl.DefaultJobExceptionHandler
+org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.DefaultJobExceptionHandler
