+++ elastic-job-lite-core/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.executor.handler.threadpool.JobExecutorServiceHandler
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.executor.handler.impl.DefaultJobExecutorServiceHandler
+org.apache.shardingsphere.elasticjob.lite.executor.handler.threadpool.impl.DefaultJobExecutorServiceHandler
