+++ elastic-job-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/executor/handler/error/impl/DefaultJobExceptionHandlerTest.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,28 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.executor.handler;
+package org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl;
 
-import java.util.concurrent.ExecutorService;
+import org.apache.shardingsphere.elasticjob.lite.event.fixture.JobEventCaller;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
 
-/**
- * Job executor service handler.
- */
-public interface JobExecutorServiceHandler {
+import static org.mockito.Mockito.atMost;
+import static org.mockito.Mockito.verify;
+
+@RunWith(MockitoJUnitRunner.class)
+public final class DefaultJobExceptionHandlerTest {
     
-    /**
-     * Create executor service.
-     * 
-     * @param jobName job name
-     * 
-     * @return executor service
-     */
-    ExecutorService createExecutorService(String jobName);
+    @Mock
+    private JobEventCaller caller;
     
-    /**
-     * Get job executor service handler type.
-     * 
-     * @return job executor service handler type
-     */
-    String getType();
+    @Test
+    public void assertHandleException() {
+        new DefaultJobExceptionHandler().handleException("test_job", new RuntimeException("test"));
+        verify(caller, atMost(1)).call();
+    }
 }
