+++ elastic-job-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/fixture/handler/ThrowJobExceptionHandler.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.elasticjob.lite.fixture.handler;
 
 import org.apache.shardingsphere.elasticjob.lite.exception.JobSystemException;
-import org.apache.shardingsphere.elasticjob.lite.executor.handler.JobExceptionHandler;
+import org.apache.shardingsphere.elasticjob.lite.executor.handler.error.JobExceptionHandler;
 
 public final class ThrowJobExceptionHandler implements JobExceptionHandler {
